rename view file
