+++ chainerui/views/result_asset.py
